# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sonoma County Map API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: markers
    description: Map marker operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /markers:
    get:
      operationId: getMarkers
      tags: [markers]
      summary: Get all markers
      description: Returns all saved map markers
      responses:
        "200":
          description: List of markers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Marker"
    post:
      operationId: createMarker
      tags: [markers]
      summary: Create a marker
      description: Add a new marker to the map
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMarkerInput"
      responses:
        "201":
          description: Created marker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Marker"
  /markers/{id}:
    get:
      operationId: getMarker
      tags: [markers]
      summary: Get a single marker
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Marker"
        "404":
          description: Not found
    put:
      operationId: updateMarker
      tags: [markers]
      summary: Update a marker
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMarkerInput"
      responses:
        "200":
          description: Updated marker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Marker"
        "404":
          description: Not found
    delete:
      operationId: deleteMarker
      tags: [markers]
      summary: Delete a marker
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /markers/stats:
    get:
      operationId: getMarkerStats
      tags: [markers]
      summary: Get marker statistics
      description: Returns counts grouped by category
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkerStats"
components:
  schemas:
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Marker:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        note:
          type: string
        category:
          type: string
          enum: [winery, restaurant]
        lat:
          type: number
        lng:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - note
        - category
        - lat
        - lng
        - createdAt
    CreateMarkerInput:
      type: object
      properties:
        name:
          type: string
        note:
          type: string
        category:
          type: string
          enum: [winery, restaurant]
        lat:
          type: number
        lng:
          type: number
      required:
        - name
        - note
        - category
        - lat
        - lng
    UpdateMarkerInput:
      type: object
      properties:
        name:
          type: string
        note:
          type: string
        category:
          type: string
          enum: [winery, restaurant]
    MarkerStats:
      type: object
      properties:
        total:
          type: integer
        wineries:
          type: integer
        restaurants:
          type: integer
      required:
        - total
        - wineries
        - restaurants