# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sonoma County Map API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all saved map markers
 * @summary Get all markers
 */
export const GetMarkersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  note: zod.string(),
  category: zod.enum(["winery", "restaurant"]),
  lat: zod.number(),
  lng: zod.number(),
  createdAt: zod.string(),
});
export const GetMarkersResponse = zod.array(GetMarkersResponseItem);

/**
 * Add a new marker to the map
 * @summary Create a marker
 */
export const CreateMarkerBody = zod.object({
  name: zod.string(),
  note: zod.string(),
  category: zod.enum(["winery", "restaurant"]),
  lat: zod.number(),
  lng: zod.number(),
});

/**
 * @summary Get a single marker
 */
export const GetMarkerParams = zod.object({
  id: zod.coerce.number(),
});

export const GetMarkerResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  note: zod.string(),
  category: zod.enum(["winery", "restaurant"]),
  lat: zod.number(),
  lng: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Update a marker
 */
export const UpdateMarkerParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMarkerBody = zod.object({
  name: zod.string().optional(),
  note: zod.string().optional(),
  category: zod.enum(["winery", "restaurant"]).optional(),
});

export const UpdateMarkerResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  note: zod.string(),
  category: zod.enum(["winery", "restaurant"]),
  lat: zod.number(),
  lng: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a marker
 */
export const DeleteMarkerParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * Returns counts grouped by category
 * @summary Get marker statistics
 */
export const GetMarkerStatsResponse = zod.object({
  total: zod.number(),
  wineries: zod.number(),
  restaurants: zod.number(),
});