# lib/api-zod/src/generated/types/openaiConversationWithMessages.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sonoma County Map API
 * OpenAPI spec version: 0.1.0
 */
import type { OpenaiMessage } from "./openaiMessage";

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: Date;
  messages: OpenaiMessage[];
}